fixup: rename Log -> Logging, move gauge package, add package docs
