+++ metric/gauge/gauge.go
