+++ metric/gauge/gauge_test.go
